requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

nltk
sentencepiece
gensim
#pytorch
torchvision
numpy
scipy
scikit-learn